Solution to Day09

+++ IntCode/ChainedMachine.cs
@@ -8,7 +8,7 @@ namespace AdventOfCode2019.Library
 {
     public class ChainedMachine : IntCodeMachine
     {
-        private static IList<Channel<int>> Inputs { get; set; }
+        private static IList<Channel<long>> Inputs { get; set; }
         private static IList<ChainedMachine> Machines { get; set; }
 
         public ChainedMachine(IntCodeMachine machine)
@@ -22,9 +22,9 @@ namespace AdventOfCode2019.Library
             Id = id;
         }
 
-        public static Channel<int> GetPopulatedChannel(IList<int> inputs)
+        public static Channel<long> GetPopulatedChannel(IList<long> inputs)
         {
-            var channel = Channel.CreateUnbounded<int>();
+            var channel = Channel.CreateUnbounded<long>();
 
             foreach (var input in inputs)
                 channel.Writer.WriteAsync(input);
@@ -32,14 +32,14 @@ namespace AdventOfCode2019.Library
             return channel;
         }
 
-        public static int ExecuteCycledProgram(IntCodeMachine machine, IList<IList<int>> inputs)
+        public static long ExecuteCycledProgram(IntCodeMachine machine, IList<IList<long>> inputs)
         {
             Inputs = inputs.Select(GetPopulatedChannel).ToList();
             Machines = Inputs.Select((_, i) => new ChainedMachine(machine, i)).ToList();
 
             Task.WaitAll(Enumerable.Range(0, Machines.Count).
                 Select(i => Task.Run(() => {
-                    Action<int> outputLambda = x => Inputs[(i + 1) % Inputs.Count].Writer.WriteAsync(x);
+                    Action<long> outputLambda = x => Inputs[(i + 1) % Inputs.Count].Writer.WriteAsync(x);
                     Machines[i].ExecuteProgram(Inputs[i], outputLambda);
                 })).
                 ToArray()
